Move snippets to test-providers.
