+++ test-apps/appui-test-app/appui-test-providers/package.json
@@ -14,7 +14,7 @@
     "build:esm": "tsc 1>&2 --module ES2020 --outDir lib/esm",
     "clean": "rimraf lib .rush/temp/package-deps*.json",
     "cover": "",
-    "docs": "",
+    "docs": "betools extract --fileExt=ts,tsx --extractFrom=./src --recursive --out=../../../generated-docs/extract",
     "lint": "eslint -f visualstudio \"./src/**/*.{ts,tsx}\" 1>&2",
     "pseudo:cjs": "betools pseudolocalize --englishDir=./src/public/locales/en --out=./lib/cjs/public/locales/en-PSEUDO",
     "pseudo:esm": "betools pseudolocalize --englishDir=./src/public/locales/en --out=./lib/esm/public/locales/en-PSEUDO",
