+++ test-apps/appui-test-app/appui-test-providers/src/docs/UiItemsProvider.Widgets.tsx
@@ -2,6 +2,7 @@
  * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
+import * as React from "react";
 // __PUBLISH_EXTRACT_START__ AppUI.UiItemsProvider.Widgets
 import {
   StagePanelLocation,
