+++ test-apps/appui-test-app/appui-test-providers/src/ui/frontstages/registerCustomFrontstage.tsx
